s/_.extend/Object.assign/g
This is personal preference rather than obvious technical superiority. They both do the same thing.

Modern JS supplies Object.assign, though, so we might as well use it. It's one less piece of custom API for an incoming developer to understand, and it saves confusion with the custom `extend` function being used elsewhere in the codebase.