Added event emitter to send updated tokens.
There were two options here:

1. create the emitter in the application and allow the consumer to use it, or
2. allow the consumer to create the emitter and inject it as a dependency

I've built it as option 1, but I'm open to discussion on this.

The creation for partner apps would look something like:

```
xeroClient = new xero.PartnerApplication(config);
    eventReceiver = xeroClient.eventEmitter;
    eventReceiver.on('xeroTokenUpdate', function(data) {
        //Store the data that was received from the xeroTokenRefresh event
        console.log("Received xero token refresh: ", data);
    });
```

This is only for partner applications, but it will be required to refresh the emitter every time a new application object is created to ensure the listening stack is shared.