Remove catch statement for event handlers
I'm not sure we should be catching and discarding exceptions from external code. If a consumer of the library causes an exception in their handler, it's up to them to deal with it. The danger with leaving the catch in is that someone hits some fatal condition and they throw an exception in order to take down the application or be handled upstream. Having it eaten inside the Xero lib would be unexpected behaviour.