Added default user agent
When the user agent was not provided in the config file, the logs were showing 'undefined'.  This has now been defaulted to the ConsumerKey for the customer to help identify this in alerting and monitoring.

We've also included a 'warn' log statement advising this may be set as mandatory in future.