Set oauthConfig.userAgent from config
Setting the UserAgent to begin with an application specific value is requested as part of Xero app certification (for partner apps).

This change ensures creation of the oauthConfig object uses the userAgent configuration property (if set in the Xero config).  This will correctly set the UserAgent header for requests made to Xero API.

If a value is not provided in the config file, it will fallback to “NodeJS-XeroAPIClient” as the userAgent prefix.

The userAgent will always be suffixed with the xero-node version and the consumerKey value from the config file.